flask/requirements.txt
africastalking==1.2.0
alembic==1.4.2
aniso8601==8.0.0
APScheduler==3.6.3
attrs==19.3.0
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
contextlib2==0.6.0.post1
Flask==1.1.2
Flask-APScheduler==1.11.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-Seeder==1.1.1
Flask-SQLAlchemy==2.4.1
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lorem==0.1.1
Mako==1.1.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
requests==2.23.0
schema==0.7.2
six==1.15.0
SQLAlchemy==1.3.17
tzlocal==2.1
urllib3==1.25.9
uWSGI==2.0.18
Werkzeug==1.0.1
zipp==3.1.0
